% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopxplorer.R
\name{associate_data_aopn}
\alias{associate_data_aopn}
\title{Associate data with the AOPN}
\usage{
associate_data_aopn(x, aopn_graph)
}
\arguments{
\item{x}{a \code{list} object of data vectors (each list element is a
different KE, or \code{vector} where the first column is the gene
symbol from the AOPN, and the subsequent columns are different conditions,
and the data within the table is some sort of response}

\item{aopn_graph}{an object of class \code{igraph} that contains the
AOPN -- a graph of the key events and their relationships.}
}
\value{
aopn_graph an object of class \code{igraph} that contains the
AOPN -- a graph of the key events and their relationships, as well as the
associated data
}
\description{
Associate data with the AOPN Key Events
}
\details{
This function associates data with the AOPN key events.
}
\examples{
\dontrun{
x <- get_aopn("Steatosis")
aopn_kes <- get_aopn_kes(x)
expression_data <- c(2, 10, 10, NA, 3, 1, 1, 3, 2, 3, 2)
associate_data_aopn(expression_data, x)
}
}

