% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopxplorer.R
\name{get_aopn}
\alias{get_aopn}
\title{Get an Adverse Outcome Pathway Network (AOPN)}
\usage{
get_aopn(ao_name, url = "http://localhost:3030/ds")
}
\arguments{
\item{ao_name}{an object of class \code{character} that is the name of the
adverse outcome of interest.}

\item{url}{a \code{character} object that is the URL to the Fuseki server. The
default is http://localhost:3030/ds}
}
\value{
aopn_graph an \code{igraph} object that contains the AOPN -- a graph
of the key events and their relationships.
}
\description{
Lists all of the key events within an AOPN
}
\details{
This function queries the AOPO for all of the key events associated with
an adverse outcome network so that the network can be built.
}
\examples{
\dontrun{
x <- get_aopn("Steatosis")
plot(x)
}
}

