% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopxplorer.R
\name{get_aopn_kes}
\alias{get_aopn_kes}
\title{Get the AOPN Key Events}
\usage{
get_aopn_kes(aopn_graph)
}
\arguments{
\item{aopn_graph}{an object of class \code{igraph} that contains the
AOPN -- a graph of the key events and their relationships.}
}
\value{
aopn_kes a \code{vector} object that contains the AOPN key events
}
\description{
Retrieve the AOPN Key Events
}
\details{
This function retrieves the AOPN key events
}
\examples{
\dontrun{
x <- get_aopn("Steatosis")
get_aopn_kes(x)
}
}

