% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopxplorer.R
\name{list_adverse_outcomes}
\alias{list_adverse_outcomes}
\title{List Adverse Outcomes}
\usage{
list_adverse_outcomes(url = "http://localhost:3030/ds")
}
\arguments{
\item{url}{a \code{character} object that is the URL to the Fuseki server. The
default is http://localhost:3030/ds}
}
\value{
adverse_outcomes a \code{vector} object that only contains the
adverse outcomes currently in the AOPO.
}
\description{
Lists all of the adverse outcomes currently in the AOPO
}
\details{
This function queries the AOPO for all of the available adverse outcomes.
}
\examples{
\dontrun{
list_adverse_outcomes()
}
}

