% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopxplorer.R
\name{mapAttributes}
\alias{mapAttributes}
\title{Maps attributes}
\usage{
mapAttributes(attr.names, all.attr, i)
}
\arguments{
\item{attr.names}{the attribute names}

\item{all.attr}{all of the attributes}

\item{i}{not sure}
}
\value{
attr
}
\description{
Maps attributes
}
\details{
Maps attributes.
}

