% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopxplorer.R
\name{send_aopn}
\alias{send_aopn}
\title{Send Adverse Outcome Pathway Network (AOPN) to Cytoscape}
\usage{
send_aopn(cytoscape_aopn, url = "http://localhost:1234/v1")
}
\arguments{
\item{cytoscape_aopn}{an object of class \code{character} that is a JSON
encoded AOPN}

\item{url}{an object of class \code{character} that is the base URL for
connecting to Cytoscape via CyREST. Default is http://localhost:1234/v1.}
}
\value{
res an result object.
}
\description{
Sends an AOPN to Cytoscape for visualization
}
\details{
This function sends an AOPN to Cytoscape for visualization
}
\examples{
\dontrun{
x <- get_aopn("Steatosis")
cytoscape_aopn <- toCytoscape(x)
send_aopn(cytoscape_aopn)
}
}

