% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopxplorer.R
\name{toCytoscape}
\alias{toCytoscape}
\title{Converts AOPN Igraph Object to CytoscapeJSON}
\usage{
toCytoscape(igraphobj)
}
\arguments{
\item{igraphobj}{an object of class \code{igraph} that is the AOPN of
interest}
}
\value{
cytoscape_aopn a \code{character} object that contains the JSON encoded AOPN
}
\description{
Converts the AOPN to CytoscapeJSON
}
\details{
This function converts an igraph network to a cytoscape JSON network
}
\examples{
\dontrun{
x <- get_aopn("Steatosis")
cytoscape_aopn <- toCytoscape(x)
}
}

