% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopxplorer.R
\name{trim_results}
\alias{trim_results}
\title{Trim Ontology Results}
\usage{
trim_results(x)
}
\arguments{
\item{x}{an object of class \code{vector} that contains ontology class
or individual names with the namespace attached.}
}
\value{
trimmed_results a \code{vector} object that only contains the class
or individual names.
}
\description{
Trims off the namespace from ontology items.
}
\details{
This function trims off the namespace information from
ontology items. It should only be used when an rdfs:label has not
been specified for the class or individual.
}
\examples{
\dontrun{
x <- "<http://aopkb.org/aop_ontology#Steatosis>"
trim_results(x)
}
}

